agent-ai/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
filelock==3.17.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
openai==1.60.2
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==70.2.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.7.0.dev20250128
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
urllib3==2.3.0
